+++ example/lib/pages/home.dart
@@ -3,6 +3,6 @@ import 'package:get_server/get_server.dart';
 class HomePage extends GetView {
   @override
   build(Context context) {
-    return context.send('Hello, you are on home');
+    return Text('Hello, you are on home');
   }
 }