+++ example/lib/pages/html.dart
@@ -6,6 +6,6 @@ class HtmlPage extends GetView {
   @override
   build(Context context) {
     final path = '${Directory.current.path}/example/web/index.html';
-    return context.sendHtml(path);
+    return Html(path);
   }
 }