+++ example/lib/pages/json.dart
@@ -3,7 +3,7 @@ import 'package:get_server/get_server.dart';
 class JsonPage extends GetView {
   @override
   build(Context context) {
-    return context.sendJson({
+    return Json({
       "fruits": ["banana", "apple", "orange"]
     });
   }