+++ example/lib/pages/socket.dart
@@ -3,7 +3,7 @@ import 'package:get_server/get_server.dart';
 class SocketPage extends GetView {
   @override
   build(Context context) {
-    context.ws.listen((socket) {
+    return Socket(context, builder: (socket) {
       socket.onMessage.listen((data) {
         print('data: $data');
         socket.send(data);