+++ example/lib/pages/user.dart
@@ -4,15 +4,6 @@ class UserPage extends GetView {
   @override
   build(Context context) {
     String name = context.param('name');
-    if (name == "pedro") {
-      // return page not found
-      return context.pageNotFound();
-
-      // you can return other status code:
-      // context
-      // ..statusCode(302)
-      // ..close();
-    }
-    return context.send('Welcome, ${context.param('name')} !');
+    return Text('Welcome, $name !');
   }
 }